PERFFORJ-58,37 : Added compile profile for ajc, upgrade to 1.6.11.
Introduces a Maven Profile "ajc" which can be activated during builds by
passing -Pajc to Maven CLI, which activates aspectj-maven-plugin. This
is indended as a temporary/testing interim for ajc-compilation rather than
a permanent distribution solution.

The AspectJ version is also pulled into a property aspectjVersion at the
same time, and version is updated to 1.6.11.  1.6.1 presents problems with
other changes on this branch when using the ajc compiler.